pyproject.toml
[project]
name = "requirements-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-anthropic>=1.3.4",
    "langchain-core>=1.2.15",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
